// packages/dashboard/src/http/types/ListEvents.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ListEventsQuery = {
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    from: string;
    /**
     * @description
     * Format: `date-time`
     * @type string
    */
    to: string;
};

export type ListEventsStatus200 = {
    data: {
        id: string;
        userId: string;
        title: string;
        description: string | null;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        startAt: string;
        isGlobal: boolean;
        reminderMinutesBefore: number | null;
        /**
         * @description
         * Format: `date-time`
         * @type string
        */
        createdAt: string;
    }[];
};

export type ListEventsOptions = {
    body?: never;
    path?: never;
    query: ListEventsQuery;
    headers?: never;
};

export type ListEventsResponses = {
    "200": ListEventsStatus200;
};

/**
 * @description Union of all possible responses
*/
export type ListEventsResponse = ListEventsStatus200;